Make Alexa the default LED pattern.
Really not liking the 4 different colors of google. Alexa does come closer to blue-isch of Mycroft. This is just a quick intermediate change as I will make it a config option, but will do that after or together with the creation of a Mycroft pattern.

But perhaps all ReSpeaker skill authors should combine everything into one skill, that just figures out which board you have....